[#64] markup - 캐로셀 수정 - border radius 적용
