+++ src/components/global/NavigationBar/NavigationBar.module.scss
@@ -25,7 +25,6 @@
       font-weight: 800;
       padding: 0.5rem;
 
-
       li {
         height: 1rem;
         box-sizing: border-box;
