+++ src/pages/HomePage/HomePage.module.scss
@@ -1,0 +1,4 @@
+@import 'styles/_palette.scss';
+
+.home-page {
+}
