+++ src/pages/index.js
@@ -1,3 +1,4 @@
+export { default as HomePage } from './HomePage';
 export { default as CategoryPage } from './CategoryPage';
 export { default as DetailPage } from './DetailPage';
 export { default as MyCartPage } from './MyCartPage';
